split flow.go to four go files
