+++ pilotage/module/flow.go
@@ -17,27 +17,36 @@ limitations under the License.
 package module
 
 import (
-	"bufio"
 	"encoding/json"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"strings"
 	"time"
 
 	. "github.com/logrusorgru/aurora"
-	homeDir "github.com/mitchellh/go-homedir"
 	"gopkg.in/yaml.v2"
-
-	apiv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
-	"k8s.io/client-go/tools/clientcmd"
-
-	"github.com/Huawei/containerops/common/utils"
 )
+
+const(
+	// Flow Run Model
+	CliRun      = "CliRun"
+	DaemonRun   = "DaemonRun"
+	DaemonStart = "DaemonStart"
+)
+
+// Flow is DevOps orchestration flow struct.
+type Flow struct {
+	Model   string   `json:"-" yaml:"-"`
+	URI     string   `json:"uri" yaml:"uri"`
+	Number  int64    `json:",omitempty" yaml:",omitempty"`
+	Title   string   `json:"title" yaml:"title"`
+	Version int64    `json:"version" yaml:"version"`
+	Tag     string   `json:"tag" yaml:"tag"`
+	Timeout int64    `json:"timeout" yaml:"timeout"`
+	Status  string   `json:"status,omitempty" yaml:"status,omitempty"`
+	Logs    []string `json:"logs,omitempty" yaml:"logs,omitempty"`
+	Stages  []Stage  `json:"stages,omitempty" yaml:"stages,omitempty"`
+}
 
 // JSON export flow data without
 func (f *Flow) JSON() ([]byte, error) {
@@ -140,229 +149,4 @@ func (f *Flow) LocalRun(verbose, timestamp bool) error {
 	return nil
 }
 
-// TODO filter the log print with different color.
-func (s *Stage) Log(log string, verbose, timestamp bool) {
-	s.Logs = append(s.Logs, fmt.Sprintf("[%s] %s", time.Now().String(), log))
 
-	if verbose == true {
-		if timestamp == true {
-			fmt.Println(Cyan(fmt.Sprintf("[%s] %s", time.Now().String(), strings.TrimSpace(log))))
-		} else {
-			fmt.Println(Cyan(log))
-		}
-	}
-}
-
-func (s *Stage) SequencingRun(verbose, timestamp bool, f *Flow) (string, error) {
-	s.Status = Running
-
-	s.Log(fmt.Sprintf("Stage [%s] status change to %s", s.Name, s.Status), false, timestamp)
-	f.Log(fmt.Sprintf("Stage [%s] status change to %s", s.Name, s.Status), verbose, timestamp)
-
-	for i, _ := range s.Actions {
-		action := &s.Actions[i]
-
-		s.Log(fmt.Sprintf("The Number [%d] action is running: %s", i, s.Title), false, timestamp)
-		f.Log(fmt.Sprintf("The Number [%d] action is running: %s", i, s.Title), verbose, timestamp)
-
-		if status, err := action.Run(verbose, timestamp, f); err != nil {
-			s.Status = Failure
-
-			s.Log(fmt.Sprintf("Action [%s] run error: %s", action.Name, err.Error()), false, timestamp)
-			f.Log(fmt.Sprintf("Action [%s] run error: %s", action.Name, err.Error()), verbose, timestamp)
-
-		} else {
-			s.Status = status
-		}
-
-		if s.Status == Failure || s.Status == Cancel {
-			break
-		}
-	}
-
-	return s.Status, nil
-}
-
-func (s *Stage) ParallelRun(verbose, timestamp bool, f *Flow) (string, error) {
-	s.Status = Running
-
-	s.Log(fmt.Sprintf("Stage [%s] status change to %s", s.Name, s.Status), false, timestamp)
-	f.Log(fmt.Sprintf("Stage [%s] status change to %s", s.Name, s.Status), verbose, timestamp)
-
-	resultChan := make(chan string)
-	defer close(resultChan)
-	count := 0
-	for i, _ := range s.Actions {
-		go func(index int) {
-			action := &s.Actions[index]
-			var tempStaus string
-
-			s.Log(fmt.Sprintf("The Number [%d] action is running: %s", index, s.Title), false, timestamp)
-			f.Log(fmt.Sprintf("The Number [%d] action is running: %s", index, s.Title), verbose, timestamp)
-
-			if status, err := action.Run(verbose, timestamp, f); err != nil {
-				tempStaus = Failure
-
-				s.Log(fmt.Sprintf("Action [%s] run error: %s", action.Name, err.Error()), false, timestamp)
-				f.Log(fmt.Sprintf("Action [%s] run error: %s", action.Name, err.Error()), verbose, timestamp)
-			} else {
-				tempStaus = status
-			}
-			resultChan <- tempStaus
-
-		}(i)
-	}
-
-	for {
-		if result, ok := <-resultChan; ok {
-			count++
-			s.Status = result
-			if result == Failure || result == Cancel || count == len(s.Actions) {
-				return s.Status, nil
-			}
-		}
-	}
-	return s.Status, nil
-}
-
-// TODO filter the log print with different color.
-func (a *Action) Log(log string, verbose, timestamp bool) {
-	a.Logs = append(a.Logs, fmt.Sprintf("[%s] %s", time.Now().String(), log))
-
-	if verbose == true {
-		if timestamp == true {
-			fmt.Println(Cyan(fmt.Sprintf("[%s] %s", time.Now().String(), strings.TrimSpace(log))))
-		} else {
-			fmt.Println(Cyan(log))
-		}
-	}
-}
-
-func (a *Action) Run(verbose, timestamp bool, f *Flow) (string, error) {
-	a.Status = Running
-
-	a.Log(fmt.Sprintf("Action [%s] status change to %s", a.Name, a.Status), false, timestamp)
-	f.Log(fmt.Sprintf("Action [%s] status change to %s", a.Name, a.Status), verbose, timestamp)
-
-	for i, _ := range a.Jobs {
-		job := &a.Jobs[i]
-
-		a.Log(fmt.Sprintf("The Number [%d] job is running: %s", i, a.Title), false, timestamp)
-		f.Log(fmt.Sprintf("The Number [%d] job is running: %s", i, a.Title), verbose, timestamp)
-
-		if status, err := job.Run(a.Name, verbose, timestamp, f); err != nil {
-			a.Status = Failure
-
-			a.Log(fmt.Sprintf("Job [%d] run error: %s", i, err.Error()), false, timestamp)
-			f.Log(fmt.Sprintf("Job [%d] run error: %s", i, err.Error()), verbose, timestamp)
-
-		} else {
-			a.Status = status
-		}
-
-		if a.Status == Failure || a.Status == Cancel {
-			break
-		}
-
-	}
-
-	return a.Status, nil
-}
-
-// TODO filter the log print with different color.
-func (j *Job) Log(log string, verbose, timestamp bool) {
-	j.Logs = append(j.Logs, fmt.Sprintf("[%s] %s", time.Now().String(), log))
-
-	if verbose == true {
-		if timestamp == true {
-			fmt.Println(Cyan(fmt.Sprintf("[%s] %s", time.Now().String(), strings.TrimSpace(log))))
-		} else {
-			fmt.Println(Cyan(log))
-		}
-	}
-}
-
-func (j *Job) Run(name string, verbose, timestamp bool, f *Flow) (string, error) {
-	home, _ := homeDir.Dir()
-
-	randomContainerName := fmt.Sprintf("%s-%s", name, utils.RandomString(10))
-
-	if config, err := clientcmd.BuildConfigFromFlags("", fmt.Sprintf("%s/.kube/config", home)); err != nil {
-		return Failure, err
-	} else {
-		if clientSet, err := kubernetes.NewForConfig(config); err != nil {
-			return Failure, err
-		} else {
-			p := clientSet.CoreV1Client.Pods(apiv1.NamespaceDefault)
-
-			if _, err := p.Create(
-				&apiv1.Pod{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "Pod",
-						APIVersion: "v1",
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Name: randomContainerName,
-					},
-					Spec: apiv1.PodSpec{
-						Containers: []apiv1.Container{
-							{
-								Name:  randomContainerName,
-								Image: j.Endpoint,
-								Env: []apiv1.EnvVar{
-									{
-										Name:  "CO_DATA",
-										Value: j.Environments[0]["CO_DATA"],
-									},
-								},
-								Resources: apiv1.ResourceRequirements{
-									Requests: apiv1.ResourceList{
-										apiv1.ResourceCPU:    resource.MustParse(j.Resources.CPU),
-										apiv1.ResourceMemory: resource.MustParse(j.Resources.Memory),
-									},
-								},
-							},
-						},
-						RestartPolicy: apiv1.RestartPolicyNever,
-					},
-				},
-			); err != nil {
-				j.Status = Failure
-				return Failure, err
-			}
-
-			j.Status = Pending
-			time.Sleep(time.Second * 5)
-
-			req := p.GetLogs(randomContainerName, &apiv1.PodLogOptions{
-				Follow:     true,
-				Timestamps: false,
-			})
-
-			if read, err := req.Stream(); err != nil {
-				// TODO Parse ContainerCreating error
-			} else {
-				reader := bufio.NewReader(read)
-				for {
-					line, err := reader.ReadString('\n')
-
-					if err != nil {
-						if err == io.EOF {
-							break
-						}
-
-						j.Status = Failure
-						return Failure, nil
-					}
-					j.Status = Running
-
-					j.Log(line, false, timestamp)
-					f.Log(line, verbose, timestamp)
-				}
-			}
-		}
-	}
-
-	j.Status = Success
-	return Success, nil
-}
